Add/correct visibility modifiers
Most things are intended to be private, but our inner classes need to be
public so they can be used by external users.

External interface functions getUriToPlay and getStartTimeCode shouldn't
be public. I've changed them to be protected so that subclasses can use
them to change the intent parsing behaviour.